refactor: 폼 관련 상수 RoomForm 으로 이동
