+++ src/RoomForm/constants/literals.ts
@@ -62,3 +62,13 @@ export const ANNOUNCEMENT = {
 
 // 새 종류 별 방 타입
 export type RoomType = (typeof ROOM_TYPES)[number];
+
+// 폼 유효성 검사 메시지
+export const FORM_MESSAGE = {
+  ANNOUNCEMENT: `공지사항은 ${ANNOUNCEMENT.min}글자에서 ${ANNOUNCEMENT.max}글자 사이여야 해요.`,
+  ROUTINE_NAME: `루틴 내용은 ${ROUTINE_NAME.min}글자에서 ${ROUTINE_NAME.max}글자 사이여야 해요.`,
+  ROOM_NAME: `방 제목은 ${ROOM_NAME.min}글자에서 ${ROOM_NAME.max}글자 사이여야 해요.`,
+  USER_COUNT: `인원을 올바르게 선택해주세요.`,
+  PASSWORD: `비밀번호는 ${PASSWORD.min}자리에서 ${PASSWORD.max}자리의 숫자여야 해요.}`,
+  ONLY_NUMERIC_PASSWORD: `비밀번호는 숫자로만 입력해주세요.`
+};
