+++ src/RoomNew/steps/RoutineStep/MemberSection.tsx
@@ -1,8 +1,8 @@
 import { useCallback } from 'react';
 import { useFormContext } from 'react-hook-form';
+import { USER_COUNT } from '@/RoomForm/constants/literals';
 import { descriptionStyle, errorStyle } from '../../constants/styles';
 import { Inputs } from '../../constants/form';
-import { USER_COUNT } from '../../constants/roomLiterals';
 import { detailSectionStyle, headingStyle, iconButtonStyle } from './styles';
 import { Icon } from '@/shared/Icon';
